requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
sqlparse==0.4.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2